% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gompertz.R
\name{evalGompertz}
\alias{evalGompertz}
\title{Evaluate Estimated Gompertz}
\usage{
evalGompertz(x, id, l = FALSE, deriv = FALSE, mc = 0, level = 0.975)
}
\arguments{
\item{x}{argument (age in centuries) vector at which to evaluate the function}

\item{id}{the estimate identifier}

\item{l}{in log-scale?}

\item{deriv}{derivate instead?}

\item{mc}{If > 1, sample posterior (approximated by mv-normal) of the parameters, evaluate function, and compute pointwise envelopes.}

\item{level}{if mc > 1, the 100xlevel\% envelopes are returned alongside the mean. Recommend using > 1000.}
}
\description{
Evaluate Estimated Gompertz
}
\details{
Note that the posterior envelope will be a bit off since the joined posterior is most likely not gaussian.
}
