% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{getCurves}
\alias{getCurves}
\title{Get Posterior Mean Curve of Particular VMI Biomass Experiment Groups in Long Format}
\usage{
getCurves(
  region = "SOUTH",
  fertility = "1-2",
  soil = "mineral",
  with_re = TRUE,
  err_model = "M",
  log_scale = TRUE,
  id
)
}
\arguments{
\item{region}{which region: SOUTH, MIDDLE, NORTH}

\item{fertility}{which fertility class: "1-2", "3", "4", "5-8"}

\item{soil}{which soil type: mineral, "drained peatl.", "undrained peatl."}

\item{with_re}{model with plot-wise random intercepts? (default: TRUE)}

\item{err_model}{default: M. Additive or Multiplicative error model?}

\item{log_scale}{return the curve in log-scale?}

\item{id}{this/these particular results by internal identifier (see \link{lookup})}
}
\description{
Get specific results. The results will be returned as a table
}
\seealso{
\link[vmiBiomass]{results}, \link[vmiBiomass]{lookup}
}
