% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{getResults}
\alias{getResults}
\title{Get Estimates of Particular VMI Biomass Experiment Groups in Long Format}
\usage{
getResults(
  zone = "SOUTH",
  fertility = "1-2",
  soil = "mineral",
  with_re = TRUE,
  err_model = "A",
  nat = FALSE,
  long = FALSE
)
}
\description{
Get specific results. The results will be returned as a tibble.
}
\seealso{
\link[vmiBiomass]{results}, \link[vmiBiomass]{lookup}
}
