% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gompertz.R
\name{gompertz}
\alias{gompertz}
\title{Gompertz Curve}
\usage{
gompertz(x, p)
}
\arguments{
\item{x}{argument}

\item{p}{parameters, as c(A,m,k) (see details).}
}
\description{
Gompertz Curve
}
\details{
The Gompertz parametric curve we used is \deqn{G(t) = A * exp(-m*k^x)}.
}
