% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gompertz.R
\name{gompertzDeriv}
\alias{gompertzDeriv}
\title{Derivative of Gompertz Curve}
\usage{
gompertzDeriv(t, p)
}
\arguments{
\item{t}{argument}

\item{p}{parameters, as c(A,m,k) (see details).}
}
\description{
Derivative of Gompertz Curve
}
\details{
The derivative in argument of the Gompertz parametric curve.
}
\seealso{
\link{gompertz}
}
