% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{m1.stan}
\alias{m1.stan}
\title{The additive error-model without RE's}
\format{
An object of class \code{character} of length 1.
}
\usage{
m1.stan
}
\description{
The additive error-model without RE's
}
\examples{
#not run
# dat1: data for one group
modA <- stan_model( model_code = m1.stan , verbose=TRUE)
fitA <- sampling(modA,
                 data =
       list( n = nrow(dat1),
             p = length(unique(dat1$group)),
             group = dat1$group, # soiltype x zone x fertility class
             ika = dat1$age,
             biomassa = dat1$biomass,
             weight = 1/sqrt(dat1$area),
             nug_var = 10),
verbose = FALSE,
cores = 4,
iter=6000)

}
\keyword{datasets}
