% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{m3re.stan}
\alias{m3re.stan}
\title{The multiplicative error-model with RE's}
\format{
An object of class \code{character} of length 1.
}
\usage{
m3re.stan
}
\description{
The multiplicative error-model with RE's
}
\seealso{
\link{m1.stan}
}
\keyword{datasets}
