% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{mA.stan}
\alias{mA.stan}
\title{The additive error-model without RE's}
\format{
An object of class \code{character} of length 1.
}
\usage{
mA.stan
}
\description{
The additive error-model without RE's
}
\examples{
#not run
# dat1: data-frame for one group
modA <- stan_model( model_code = mA.stan , verbose=TRUE)
fitA <- sampling(modA,
                 data =
       list( n = nrow(dat1),
             p = length(unique(dat1$group)),
             group = dat1$group, # soiltype x zone x fertility class
             age = dat1$age,
             biomass = dat1$biomass,
             weight = dat1$weight/sum(dat1$weight) * nrow(dat1),
             nug_var = 10),
verbose = FALSE,
cores = 4,
iter=6000)

}
\keyword{datasets}
