% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{mAre.stan}
\alias{mAre.stan}
\title{The additive error-model with RE's}
\format{
An object of class \code{character} of length 1.
}
\usage{
mAre.stan
}
\description{
The additive error-model with RE's
}
\seealso{
\link{mA.stan}
}
\keyword{datasets}
