% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{mM.stan}
\alias{mM.stan}
\title{The multiplicative error-model without RE's}
\format{
An object of class \code{character} of length 1.
}
\usage{
mM.stan
}
\description{
The multiplicative error-model without RE's
}
\seealso{
\link{mA.stan}
}
\keyword{datasets}
