% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-models.R
\docType{data}
\name{mMre.stan}
\alias{mMre.stan}
\title{The multiplicative error-model with RE's}
\format{
An object of class \code{character} of length 1.
}
\usage{
mMre.stan
}
\description{
The multiplicative error-model with RE's
}
\seealso{
\link{mM.stan}
}
\keyword{datasets}
