% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\docType{data}
\name{results}
\alias{results}
\title{Results Object}
\format{
a list with sublists by a unique id, see \link[vmiBiomass]{lookup}. The structure for each element is
a list of posterior statistics:
\describe{
\item{ranef}{(optional) the group-wise estimated random effects}
\item{env}{95\% posterior envelope for the curve as a matrix, first column gives the argument}
\item{lenv}{envelope for the log-curve (base e)}
\item{denv}{envelope for the derivative}
\item{dlenv}{envelope for the derivative of the log-curve}
\item{est}{Posterior statistics for the parameters, a list(stat = table, Cov = matrix)}
\item{estnat}{ Posterior statistics for derived "interpretable parameters"}
}
}
\usage{
results
}
\description{
Results for the biomass curve fitting. Gompertz curve.
}
\details{
Note: the model was fitted with age in centuries.
}
\seealso{
\link[vmiBiomass]{getResults}, \link[vmiBiomass]{lookup}
}
\keyword{datasets}
