% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmiBiomass-package.R
\docType{package}
\name{vmiBiomass-package}
\alias{vmiBiomass}
\alias{vmiBiomass-package}
\title{vmiBiomass: Supplemental Package for Repo et. al. 2021 VMI Biomass Paper}
\description{
Model fitting scripts and result objects for the paper
 Repo et al. "Age-dependence of stand biomass in managed boreal forests 
 abased on the Finnish National Forest Inventory data"
}
\details{
See \link{results} for the fitted parameters,
\link{mA.stan}, \link{mAre.stan}, \link{mM.stan} and \link{mMre.stan} for the model fitting,
and \link{evalGompertz} for drawing curves from the posteriors.

Example usage provided in the vignette, see:
\code{vignette("basics", package="vmiBiomass")} .
}
\keyword{internal}
